refactor(play): extract InGameListeners components (S26.0b)
Deuxieme tranche du refactor S26.0 :
play/[id]/page.tsx 1557 -> 1514 lignes (-43, total -152 depuis 1666).

- Extrait 3 helpers UI in-game (TurnNotificationListener,
  SoundEffectsListener, SoundToggleButton) dans
  apps/web/app/play/[id]/components/InGameListeners.tsx avec
  interfaces props nommees.
- page.tsx perd 3 imports inutilises (useTurnNotification,
  useSoundEffects, getSoundManager).

Slices restantes pour S26.0 : SetupPhaseUI (~280 lignes selon les
references setup* dans page.tsx), BlockChoiceFlow, ThrowTeamMateFlow,
puis types Move unions pour eliminer les `as any` (13 occurrences).
